requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.12
django-cors-headers==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.6
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.4
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
